+++ OtripleS.Portal.Web.Tests.Unit/Services/Students/StudentServiceTests.Exceptions.cs
@@ -16,28 +16,6 @@ namespace OtripleS.Portal.Web.Tests.Unit.Services.Students
 {
     public partial class StudentServiceTests
     {
-        public static TheoryData ValidationApiExceptions()
-        {
-            string exceptionMessage = GetRandomString();
-            var responseMessage = new HttpResponseMessage();
-
-            var httpResponseBadRequestException =
-                new HttpResponseBadRequestException(
-                    responseMessage: responseMessage,
-                    message: exceptionMessage);
-
-            var httpResponseConflictException =
-                new HttpResponseConflictException(
-                    responseMessage: responseMessage,
-                    message: exceptionMessage);
-
-            return new TheoryData<Exception>
-            {
-                httpResponseBadRequestException,
-                httpResponseConflictException
-            };
-        }
-
         [Theory]
         [MemberData(nameof(ValidationApiExceptions))]
         public async Task ShouldThrowDependencyValidationExceptionOnRegisterIfBadRequestErrorOccursAndLogItAsync(
@@ -74,33 +52,7 @@ namespace OtripleS.Portal.Web.Tests.Unit.Services.Students
             this.apiBrokerMock.VerifyNoOtherCalls();
             this.loggingBrokerMock.VerifyNoOtherCalls();
         }
-
-        public static TheoryData CriticalApiExceptions()
-        {
-            string exceptionMessage = GetRandomString();
-            var responseMessage = new HttpResponseMessage();
-
-            var httpRequestException =
-                new HttpRequestException();
-
-            var httpResponseUrlNotFoundException =
-                new HttpResponseUrlNotFoundException(
-                    responseMessage: responseMessage,
-                    message: exceptionMessage);
-
-            var httpResponseUnAuthorizedException =
-                new HttpResponseUnauthorizedException(
-                    responseMessage: responseMessage,
-                    message: exceptionMessage);
-
-            return new TheoryData<Exception>
-            {
-                httpRequestException,
-                httpResponseUrlNotFoundException,
-                httpResponseUnAuthorizedException
-            };
-        }
-
+ 
         [Theory]
         [MemberData(nameof(CriticalApiExceptions))]
         public async Task ShouldThrowCriticalDependencyExceptionOnRegisterIfUrlNotFoundErrorOccursAndLogItAsync(
@@ -136,29 +88,6 @@ namespace OtripleS.Portal.Web.Tests.Unit.Services.Students
 
             this.apiBrokerMock.VerifyNoOtherCalls();
             this.loggingBrokerMock.VerifyNoOtherCalls();
-        }
-
-
-        public static TheoryData DependencyApiExceptions()
-        {
-            string exceptionMessage = GetRandomString();
-            var responseMessage = new HttpResponseMessage();
-
-            var httpResponseException =
-                new HttpResponseException(
-                    httpResponseMessage: responseMessage,
-                    message: exceptionMessage);
-
-            var httpResponseInternalServerErrorException =
-                new HttpResponseInternalServerErrorException(
-                    responseMessage: responseMessage,
-                    message: exceptionMessage);
-
-            return new TheoryData<Exception>
-            {
-                httpResponseException,
-                httpResponseInternalServerErrorException
-            };
         }
 
         [Theory]
